+++ cmd/holotreeHash.go
@@ -1,11 +1,18 @@
 package cmd
 
 import (
+	"fmt"
+
 	"github.com/robocorp/rcc/common"
 	"github.com/robocorp/rcc/htfs"
+	"github.com/robocorp/rcc/operations"
 	"github.com/robocorp/rcc/pretty"
 
 	"github.com/spf13/cobra"
+)
+
+var (
+	holotreeShowBlueprint bool
 )
 
 var holotreeHashCmd = &cobra.Command{
@@ -20,9 +27,29 @@ var holotreeHashCmd = &cobra.Command{
 		_, holotreeBlueprint, err := htfs.ComposeFinalBlueprint(args, "", common.DevDependencies)
 		pretty.Guard(err == nil, 1, "Blueprint calculation failed: %v", err)
 		hash := common.BlueprintHash(holotreeBlueprint)
+
 		common.Log("Blueprint hash for %v is %v.", args, hash)
-		if common.Silent() {
-			common.Stdout("%s\n", hash)
+		if holotreeShowBlueprint {
+			common.Log("Blueprint:\n%s", holotreeBlueprint)
+		}
+
+		if holotreeJson {
+			result := make(map[string]interface{})
+			result["hash"] = hash
+			if holotreeShowBlueprint {
+				result["blueprint"] = string(holotreeBlueprint)
+			}
+
+			out, err := operations.NiceJsonOutput(result)
+			pretty.Guard(err == nil, 6, "%s", err)
+			fmt.Println(out)
+		} else {
+			if common.Silent() {
+				common.Stdout("%s\n", hash)
+				if holotreeShowBlueprint {
+					common.Stdout("Blueprint:\n%s", holotreeBlueprint)
+				}
+			}
 		}
 	},
 }
@@ -30,4 +57,6 @@ var holotreeHashCmd = &cobra.Command{
 func init() {
 	holotreeCmd.AddCommand(holotreeHashCmd)
 	holotreeHashCmd.Flags().BoolVarP(&common.DevDependencies, "devdeps", "", false, "Include dev-dependencies from the `package.yaml` when calculating the hash (only valid when dealing with a `package.yaml` file).")
+	holotreeHashCmd.Flags().BoolVarP(&holotreeJson, "json", "j", false, "Show environment as JSON.")
+	holotreeHashCmd.Flags().BoolVarP(&holotreeShowBlueprint, "show-blueprint", "", false, "Show full blueprint, not just hash.")
 }